Fix provider list
