+++ packages/esm-billing-app/src/claims/utils.ts
@@ -1,0 +1,7 @@
+export const statusColors = {
+  ENTERED: 'blue',
+  ERRORED: 'red',
+  REJECTED: 'red',
+  CHECKED: 'green',
+  VALUATED: 'purple',
+};
